[PROPOSAL] handler/signature: Handle signature verification lazily.
This is a proposal API change to PutObject() and PutObjectPart()
which takes an additional argument for verifying payload.

This approach is choosen in a way to avoid the complexities involved
in using io.Pipe() and starting a go-routine inside a handler.

This approach avoids races which currently exist in PutObject and PutObjectPart
handler.

Thanks to Krishna for his inputs on this.

Fixes #2058